# client/protoapi/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/news.service.v1.News/Ping": {
            "get": {
                "summary": "/news.service.v1.News/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "news.service.v1.News"
                ]
            }
        },
        "/news/all": {
            "get": {
                "summary": "/news/all",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.news.service.v1.NewsResp"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "news.service.v1.News"
                ]
            }
        },
        "/news/news-info": {
            "get": {
                "summary": "/news/news-info",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.news.service.v1.NewsInfoResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "news.service.v1.News"
                ]
            }
        }
    },
    "definitions": {
        ".google.protobuf.Empty": {
            "type": "object",
            "properties": {},
            "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }\n\n The JSON representation for `Empty` is empty JSON object `{}`."
        },
        ".news.service.v1.IdReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            },
            "required": [
                "id"
            ]
        },
        ".news.service.v1.NewsInfoResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        ".news.service.v1.NewsResp": {
            "type": "object",
            "properties": {
                "news": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/.news.service.v1.NewsResp.Info"
                    }
                }
            }
        },
        ".news.service.v1.NewsResp.Info": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        }
    }
}